requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
datefinder==0.7.3
Deprecated==1.2.14
deprecation==2.1.0
filigran-sseclient==1.0.1
idna==3.10
importlib-metadata==6.11.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
packaging==24.1
pika==1.3.2
prometheus_client==0.20.0
pyaes==1.6.1
pycti==6.3.2
pydantic==2.8.2
pydantic_core==2.20.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-magic==0.4.27
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
setuptools==71.1.0
simplejson==3.19.3
six==1.16.0
soupsieve==2.6
stix2==3.0.1
stix2-patterns==2.0.0
TgCrypto==1.2.5
typing_extensions==4.12.2
urllib3==2.2.3
wrapt==1.16.0
zipp==3.20.2
